ClI Guessing Game (testing random and sys module)
For this game I wanted to implement a random number guessing game into a CLI format to test the sys.argv function to better understand it's use. 3 arguments are passed: a lower range and upper range for the randint method from the random module, as well as the users guess. This makes repeating the game laborious and not as user friendly. If I have time in the future, I will try to develop a more convenient implementation.